Apply

+++ DAL/ApplyRepo.cs
@@ -1,8 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace DAL
 {
@@ -20,9 +17,20 @@ namespace DAL
             return context.Applies.Take(limit).ToList();
         }
 
-        public static List<Apply> GetAppliesByEmployee(int employeeId, int limit)
+        public static List<Apply> GetAppliesByUser(int userId, int limit)
         {
-            return context.Applies.Where(a => a.EmployeeId == employeeId).Take(limit).ToList();
+            var data = context.Employees.Where(e => e.UserId == userId).FirstOrDefault();
+            if (data == null)
+            {
+                return null;
+            }
+
+            return context.Applies.Where(a => a.EmployeeId == data.EmployeeId).Take(limit).ToList();
+        }
+
+        public static List<Apply> GetAppliesByJobPost(int jobPostId, int limit)
+        {
+            return context.Applies.Where(a => a.JobPostId == jobPostId).Take(limit).ToList();
         }
 
         public static Apply GetApply(int id)
@@ -30,19 +38,50 @@ namespace DAL
             return context.Applies.Find(id);
         }
 
-        public static void CreateApply(Apply a)
+        public static bool CreateApply(int userId, int jobPostId, Apply ap)
         {
-            context.Applies.Add(a);
+            var data = context.Employees.Where(e => e.UserId == userId).FirstOrDefault();
+            if (data == null)
+            {
+                return false;
+            }
+
+            ap.EmployeeId = data.EmployeeId;
+            ap.JobPostId = jobPostId;
+
+            context.Applies.Add(ap);
             context.SaveChanges();
+
+            return true;
         }
 
-        // TODO: Update Apply
+        public static bool UpdateApply(int applyId, Apply ap)
+        {
+            var apply = context.Applies.Find(applyId);
+            if (apply == null)
+            {
+                return false;
+            }
 
-        public static void DeleteApply(int id)
+            apply.ApplyId = ap.ApplyId;
+            context.Entry(apply).CurrentValues.SetValues(ap);
+            context.SaveChanges();
+
+            return true;
+        }
+
+        public static bool DeleteApply(int id)
         {
             var apply = context.Applies.Find(id);
+            if (apply == null)
+            {
+                return false;
+            }
+
             context.Applies.Remove(apply);
             context.SaveChanges();
+
+            return true;
         }
     }
 }
